+++ repos/paella-core/changelog.md
@@ -11,7 +11,11 @@
     * New feature: chapters. See documentation on video manifest for more information.
 - 2.4.1:    
     * Adjust the direction of the buttons in the menu and add a right chevron icon to the PopUpButton plugins.
-    * Add aria-label to back button in popups
-- 2.4.2:
+    * Add aria-label to back button in popups.
+    * New CSS variables added to customize menu items.
+- 2.5.0:
+    * New API: timeline images
+    * Improvements and bug fixes in video container message API.
     * Changes in typescript types (Plugins)
+